+++ tools/file.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 from typing import Union
-
+import base64
 from loguru import logger
 from natsort import natsorted
 
@@ -21,6 +21,23 @@ VIDEO_EXTENSIONS = {
     ".mp4",
     ".avi",
 }
+
+
+def audio_to_base64(file_path):
+    if not file_path or not Path(file_path).exists():
+        return None
+    with open(file_path, "rb") as wav_file:
+        wav_content = wav_file.read()
+        base64_encoded = base64.b64encode(wav_content)
+        return base64_encoded.decode("utf-8")
+
+
+def read_ref_text(ref_text):
+    path = Path(ref_text)
+    if path.exists() and path.is_file():
+        with path.open("r", encoding="utf-8") as file:
+            return file.read()
+    return ref_text
 
 
 def list_files(
